update react-transition-group and react-textfit
