+++ package.json
@@ -27,7 +27,7 @@
     "react-select-search": "^4.1.6",
     "react-textfit": "^1.1.1",
     "react-toggle": "^4.1.2",
-    "react-transition-group": "^1.2.1",
+    "react-transition-group": "^4.4.5",
     "react-vega": "^7.4.4",
     "react-virtualized": "^9.22.3",
     "redux": "^4.1.2",
@@ -52,6 +52,8 @@
     "@types/react-dom": "^17.0.11",
     "@types/react-helmet": "^6.1.6",
     "@types/react-test-renderer": "^16.9.5",
+    "@types/react-textfit": "^1.1.0",
+    "@types/react-transition-group": "^4.4.5",
     "@types/react-virtualized": "^9.21.21",
     "@wojtekmaj/enzyme-adapter-react-17": "^0.6.5",
     "archiver": "^5.3.1",
