+++ src/interface/utils/useWaSelector.ts
@@ -1,4 +1,4 @@
-import { RootState } from 'interface/reducers';
 import { TypedUseSelectorHook, useSelector } from 'react-redux';
+import { RootState } from 'store';
 
 export const useWaSelector: TypedUseSelectorHook<RootState> = useSelector;
